feat: 部门、岗位功能开发

+++ antweb2/src/pages/System/Post/index.tsx
@@ -1,62 +1,32 @@
-import { Button, Divider, Space, Drawer, message } from 'antd';
+import {
+  Button,
+  Divider,
+  Space,
+  message,
+  Tooltip,
+  Dropdown,
+  Popconfirm,
+  Modal,
+} from 'antd';
+import {
+  ExportOutlined,
+  EllipsisOutlined,
+  DeleteOutlined,
+  EditOutlined,
+  PlusOutlined,
+} from '@ant-design/icons';
 import {
   ActionType,
   FooterToolbar,
   PageContainer,
-  ProDescriptions,
   ProTable,
   ProColumns,
 } from '@ant-design/pro-components';
 import React, { useRef, useState } from 'react';
-import { queryPostPage } from '@/services/post';
-import CreateForm from './components/CreateForm';
-import UpdateForm, { FormValueType } from './components/UpdateForm';
-
-/**
- * 添加节点
- * @param fields
- */
-const handleAdd = async (fields: API.UserInfo) => {
-  const hide = message.loading('正在添加');
-  try {
-    await addUser({ ...fields });
-    hide();
-    message.success('添加成功');
-    return true;
-  } catch (error) {
-    hide();
-    message.error('添加失败请重试！');
-    return false;
-  }
-};
-
-/**
- * 更新节点
- * @param fields
- */
-const handleUpdate = async (fields: FormValueType) => {
-  const hide = message.loading('正在配置');
-  try {
-    await modifyUser(
-      {
-        userId: fields.id || '',
-      },
-      {
-        name: fields.name || '',
-        nickName: fields.nickName || '',
-        email: fields.email || '',
-      },
-    );
-    hide();
-
-    message.success('配置成功');
-    return true;
-  } catch (error) {
-    hide();
-    message.error('配置失败请重试！');
-    return false;
-  }
-};
+import { queryPostPage, deletePost } from '@/services/post';
+import { queryDictsByType } from '@/services/dict';
+import CreatePostForm from './components/CreatePostForm';
+import UpdatePostForm from './components/UpdatePostForm';
 
 /**
  *  删除节点
@@ -66,79 +36,134 @@ const handleRemove = async (selectedRows: API.UserInfo[]) => {
   const hide = message.loading('正在删除');
   if (!selectedRows) return true;
   try {
-    await deleteUser({
-      userId: selectedRows.find((row) => row.id)?.id || '',
-    });
+    await deletePost(selectedRows.map((row) => row.postId).join(','));
     hide();
-    message.success('删除成功，即将刷新');
+    message.success('删除成功');
     return true;
-  } catch (error) {
+  } catch {
     hide();
-    message.error('删除失败，请重试');
+    message.error('删除失败请重试!');
     return false;
   }
 };
 
 export const Component: React.FC<unknown> = () => {
-  const [createModalVisible, handleModalVisible] = useState<boolean>(false);
-  const [updateModalVisible, handleUpdateModalVisible] =
-    useState<boolean>(false);
-  const [stepFormValues, setStepFormValues] = useState({});
   const actionRef = useRef<ActionType>();
-  const [row, setRow] = useState<API.UserInfo>();
   const [selectedRowsState, setSelectedRows] = useState<API.UserInfo[]>([]);
+
   const columns: ProColumns[] = [
     {
       title: '岗位编号',
       dataIndex: 'postId',
       hideInSearch: true,
-    },
-    {
-      title: '岗位编码',
-      dataIndex: 'postCode',
-      valueType: 'text',
+      hideInForm: true,
     },
     {
       title: '岗位名称',
       dataIndex: 'postName',
       valueType: 'text',
+      width: 160,
+      hideInTable: true,
+      hideInSearch: true,
+      formItemProps: {
+        rules: [{ required: true, message: '请输入岗位名称' }],
+      },
+      colProps: { span: 12 },
+    },
+    {
+      title: '岗位编码',
+      dataIndex: 'postCode',
+      valueType: 'text',
+      width: 160,
+      formItemProps: {
+        rules: [{ required: true, message: '请输入岗位名称' }],
+      },
+      colProps: { span: 12 },
+    },
+    {
+      title: '岗位名称',
+      dataIndex: 'postName',
+      valueType: 'text',
+      width: 160,
+      hideInForm: true,
     },
     {
       title: '排序',
       dataIndex: 'postSort',
-      valueType: 'text',
+      valueType: 'digit',
+      width: 120,
+      initialValue: 0,
       hideInSearch: true,
+      formItemProps: {
+        rules: [{ required: true, message: '请输入岗位名称' }],
+      },
+      colProps: { span: 12 },
     },
     {
       title: '状态',
       dataIndex: 'status',
-      valueType: 'select',
-      valueEnum: {
-        0: { text: '正常', status: 'MALE' },
-        1: { text: '停用', status: 'FEMALE' },
+      valueType: 'radio',
+      width: 120,
+      initialValue: '0',
+      request: async () => {
+        const res = await queryDictsByType('sys_normal_disable');
+        return res.data.map((dict) => ({
+          label: dict.dictLabel,
+          value: dict.dictValue,
+        }));
       },
+      colProps: { span: 12 },
+    },
+    {
+      title: '备注',
+      dataIndex: 'remark',
+      valueType: 'textarea',
+      hideInTable: true,
+      hideInSearch: true,
+      colProps: { span: 24 },
     },
     {
       title: '创建时间',
       dataIndex: 'createTime',
-      valueType: 'dateRange',
+      valueType: 'dateTime',
+      width: 220,
+      hideInForm: true,
     },
     {
       title: '操作',
-      width: 180,
+      width: 100,
       dataIndex: 'option',
       valueType: 'option',
       render: (_, record) => (
-        <Space direction="horizontal" split={<Divider type="vertical" />}>
-          <a
-            onClick={() => {
-              handleUpdateModalVisible(true);
-              setStepFormValues(record);
-            }}
-          >
-            修改
-          </a>
-          <a href="">删除</a>
+        <Space
+          direction="horizontal"
+          split={<Divider type="vertical" />}
+          size={2}
+        >
+          <Tooltip title="修改">
+            <UpdatePostForm
+              values={record}
+              columns={columns}
+              trigger={
+                <Button type="link" size="small" icon={<EditOutlined />} />
+              }
+              onFinish={() => {
+                actionRef.current?.reload();
+              }}
+            />
+          </Tooltip>
+          <Tooltip title="删除">
+            <Popconfirm
+              title="删除记录"
+              description="您确定要删除此记录吗？"
+              onConfirm={async () => {
+                await handleRemove([record]);
+                actionRef.current?.reloadAndRest?.();
+              }}
+            >
+              <Button type="link" size="small" icon={<DeleteOutlined />} />
+            </Popconfirm>
+          </Tooltip>
         </Space>
       ),
     },
@@ -155,16 +180,34 @@ export const Component: React.FC<unknown> = () => {
         actionRef={actionRef}
         rowKey="postId"
         toolBarRender={() => [
-          <Button
-            key="add"
-            type="primary"
-            onClick={() => handleModalVisible(true)}
+          <CreatePostForm
+            columns={columns}
+            trigger={
+              <Button type="primary" icon={<PlusOutlined />} key="add">
+                新建
+              </Button>
+            }
+            onFinish={() => {
+              actionRef.current?.reload();
+            }}
+          />,
+
+          <Dropdown
+            key="menu"
+            menu={{
+              items: [
+                {
+                  label: '导出',
+                  icon: <ExportOutlined />,
+                  key: 'export',
+                },
+              ],
+            }}
           >
-            新建
-          </Button>,
-          <Button key="export" onClick={() => handleModalVisible(true)}>
-            导出
-          </Button>,
+            <Button>
+              <EllipsisOutlined />
+            </Button>
+          </Dropdown>,
         ]}
         request={async (params, sorter, filter) => {
           const { code, rows, total } = await queryPostPage({
@@ -201,79 +244,26 @@ export const Component: React.FC<unknown> = () => {
         >
           <Button
             onClick={async () => {
-              await handleRemove(selectedRowsState);
-              setSelectedRows([]);
-              actionRef.current?.reloadAndRest?.();
+              Modal.confirm({
+                title: '删除记录',
+                content: '您确定要删除选中的记录吗？',
+                onOk: async () => {
+                  const ok = await handleRemove(selectedRowsState);
+                  if (ok) {
+                    setSelectedRows([]);
+                    actionRef.current?.reloadAndRest?.();
+                    Promise.resolve();
+                  } else {
+                    Promise.reject();
+                  }
+                },
+              });
             }}
           >
             批量删除
           </Button>
-          <Button type="primary">批量审批</Button>
         </FooterToolbar>
       )}
-
-      <CreateForm
-        onCancel={() => handleModalVisible(false)}
-        modalVisible={createModalVisible}
-      >
-        <ProTable
-          onSubmit={async (value) => {
-            const success = await handleAdd(value);
-            if (success) {
-              handleModalVisible(false);
-              if (actionRef.current) {
-                actionRef.current.reload();
-              }
-            }
-          }}
-          rowKey="id"
-          type="form"
-          columns={columns}
-        />
-      </CreateForm>
-      {stepFormValues && Object.keys(stepFormValues).length ? (
-        <UpdateForm
-          onSubmit={async (value) => {
-            const success = await handleUpdate(value);
-            if (success) {
-              handleUpdateModalVisible(false);
-              setStepFormValues({});
-              if (actionRef.current) {
-                actionRef.current.reload();
-              }
-            }
-          }}
-          onCancel={() => {
-            handleUpdateModalVisible(false);
-            setStepFormValues({});
-          }}
-          updateModalVisible={updateModalVisible}
-          values={stepFormValues}
-        />
-      ) : null}
-
-      <Drawer
-        width={600}
-        open={!!row}
-        onClose={() => {
-          setRow(undefined);
-        }}
-        closable={false}
-      >
-        {row?.name && (
-          <ProDescriptions<API.UserInfo>
-            column={2}
-            title={row?.name}
-            request={async () => ({
-              data: row || {},
-            })}
-            params={{
-              id: row?.name,
-            }}
-            columns={columns}
-          />
-        )}
-      </Drawer>
     </PageContainer>
   );
 };
